% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abandon_Rate_Gas.r
\name{Abandon_Rate_Gas}
\alias{Abandon_Rate_Gas}
\title{Returns an estimate of a gas well's abandonment rate (Mcf/d) based on the depth of the completed interval (m).}
\usage{
Abandon_Rate_Gas(mTVD)
}
\arguments{
\item{mTVD}{depth of completed interval}
}
\description{
\code{Abandon_Rate_Gas}
}
\examples{
depth <- 953.2
Abandon_Rate_Gas(depth)
}
