% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abandon_Rate_Oil.r
\name{Abandon_Rate_Oil}
\alias{Abandon_Rate_Oil}
\title{Returns an estimate of an oil well's abandonment rate (bbl/d) based on the depth of the completed interval (m)}
\usage{
Abandon_Rate_Oil(mTVD)
}
\arguments{
\item{mTVD}{depth of completed interval}
}
\description{
\code{Abandon_Rate_Gas}
}
\examples{
depth <- 953.2
Abandon_Rate_Oil(depth)
}
