% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Boot.Harm.Reserves.r
\name{Boot.Harm.Reserves}
\alias{Boot.Harm.Reserves}
\title{Calculates harmonic reserves with uncertainty using bootstrap method}
\usage{
Boot.Harm.Reserves(x, iters)
}
\arguments{
\item{x}{data frame to parse}

\item{iters}{number of iteration to use for bootstrap estiamtes}
}
\description{
\code{Boot.Harm.Reserves}
All iterations that give a cumulative production less than the actual cum are rejected and a replacement iteration is added
}
\examples{
data <- data.frame(q=seq(500,5,-5),Q=seq(20,2000,20),qf=rep(25,100))
iters <- 5
Boot.Harm.Reserves(data, iters)
}
