% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.lqs.r
\name{boot.lqs}
\alias{boot.lqs}
\title{Returns a function that does a quantile regression linear model to a random sample of a dataset... ie do a bootstrap fit}
\usage{
boot.lqs(formula, data, ...)
}
\arguments{
\item{formula}{linear model of form "formula", generally like q~Q}

\item{data}{data frame to parse}

\item{...}{additional paramaters to pass on}
}
\description{
\code{boot.lqs}
}
\examples{
formula <- q~Q
data <- data.frame(q=5*(100:1),Q=20*(1:100),qf=rep(25,100))
boot.lqs(formula, data)
}
