% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.lm.full.r
\name{break.lm.full}
\alias{break.lm.full}
\title{Returns the full analysis of breakpoints (no more than 3) for a linear model of form "formula" on data "x".
To install the breakpoints packages \code{install.packages("breakpoint",repos="http://cran.r-project.org/", type="source")}}
\usage{
break.lm.full(formula, x)
}
\arguments{
\item{formula}{linear model of form "formula", generally like q~Q}

\item{x}{data frame to parse}
}
\description{
\code{break.lm.full}
}
\examples{
formula <- q~Q
data <- data.frame(q=seq(500,5,-5),Q=seq(20,2000,20),qf=rep(25,100))
#To show the fitted regression lines, which are discontinuous, ...
#require(strucchange)
#bp.seat <- break.lm.full(formula, data)
#plot(data$Q,data$q)
#lines(data$Q,fitted(lm(q ~ breakfactor(bp.seat)/Q - 1, data = data)),col="red")
}
