% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.lm.r
\name{break.lm}
\alias{break.lm}
\title{Returns the index of the optimum breakpoints (no more than 3) for a linear model of form "formula" on data "x".}
\usage{
break.lm(formula, x)
}
\arguments{
\item{formula}{linear model of form "formula", generally like q~Q}

\item{x}{data frame to parse}
}
\description{
\code{break.lm}
}
\examples{
formula <- q~Q
data <- data.frame(q=seq(500,5,-5),Q=seq(20,2000,20),qf=rep(25,100))
break.lm(formula, data)
}
