% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gas_Data.r
\name{Gas_Data}
\alias{Gas_Data}
\title{Returns a data frame with just the variables we need to perform a gas decline.}
\usage{
Gas_Data(df)
}
\arguments{
\item{df}{data frame to parse}
}
\description{
\code{Gas_Data}
}
\examples{
data <- data.frame(PROD_DATE=50,COND_CAL_DAY=5,OIL_CAL_DAY=12,CUM_OIL_BT=50,GAS_CAL_DAY=50,CUM_GAS=200,X_TD_TVD=200)
Gas_Data(data)
}
