% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oil_Data.r
\name{Oil_Data}
\alias{Oil_Data}
\title{Returns a data frame with just the variables we need to perform an oil decline.}
\usage{
Oil_Data(df)
}
\arguments{
\item{df}{data frame to parse}
}
\description{
\code{Oil_Data}
}
\examples{
data <- data.frame(PROD_DATE=50,COND_CAL_DAY=5,OIL_CAL_DAY=12,CUM_OIL_BT=50,GAS_CAL_DAY=50,CUM_GAS=200,X_TD_TVD=200)
Oil_Data(data)
}
