% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Peak_Rate.r
\name{Peak_Rate}
\alias{Peak_Rate}
\title{Returns a data frame with the UWI and peak monthly gas rate.}
\usage{
Peak_Rate(x)
}
\arguments{
\item{x}{data frame to parse}
}
\value{
data frame containing the UWI and peak monthly gas rate of \code{x}
}
\description{
\code{Peak_Rate}
}
\examples{
data <- data.frame(X_UWI_DISPLAY="Well", GAS_CAL_DAY=5*(10:1))
Peak_Rate(data)
}
